Split repository.ts: extract admin stats, reports, answer feedback (13/N, final)
Thirteenth and final slice of the repository.ts carve-up (audit L14). Moves the
last three domains, 1,516 lines:

- repository/adminStats.ts (345) — super-admin observability: usage/cost
  aggregates, per-model and per-platform breakdowns, admin-activity summaries,
  and the background-job cost rollup. The two source sections move together
  because they are MUTUALLY coupled (usageStats folds in sumBackgroundJobCosts;
  the job-cost section reads usageStats' window shape), so splitting them would
  need cross-module imports in both directions — a cycle.
- repository/contentReports.ts (594) — member-facing abuse/spam reports: intake,
  admin triage, resolution, repeat-target warning.
- repository/answerFeedback.ts (577) — rate_answer signal (#118), its daily cap,
  unhelpful clustering, and the low-rated-knowledge rollups.

🔒 SCOPING VERIFICATION (contentReports is the most scoping-dense domain in the
repo; answerFeedback and adminStats also carry scoped reads):
- 15 scoped functions moved; every `conversationIds` parameter form
  (`readonly string[]`, `readonly string[] | null`, `?: readonly string[]`)
  present verbatim in origin/main;
- all 3 `conversation_id = ANY($n)` predicates present verbatim;
- no predicate reformatted, no parameter dropped.
Each module header states the scoping contract.

All 40 moved symbols byte-identical against origin/main.

repository.ts is now 570 lines — down from 7,137 when this campaign started, a
92% reduction — and is now just the interactions/memory core plus the re-export
barrel. Two things deliberately remain: the interactions/recall functions
themselves (the natural core of the module) and the member-facing weekly digest
guard, which uses pageKeyOf from context/docsIngest.ts and would route that
file's pre-existing import cycle through a submodule.

Every one of the 26 files under src/storage/repository/ is reachable through the
unchanged repository.ts barrel, so all ~30 importing files and every test are
untouched across the whole campaign.
