+++ src/storage/repository.ts
@@ -34,13 +34,7 @@ export * from './repository/shared.js';
 export * from './repository/devTeamWatches.js';
 // `export *` does not bind names into this module's own scope, so anything
 // still living here that uses an extracted symbol must import it explicitly.
-import {
-  cosineSim,
-  invalidateDigestsForInteractions,
-  QUESTION_CLUSTER_SIMILARITY_THRESHOLD,
-} from './repository/shared.js';
-import { AUTO_ENROLL_ACTOR } from './repository/members.js';
-import { sumShortcutHits } from './repository/shortcutHits.js';
+import { invalidateDigestsForInteractions } from './repository/shared.js';
 export * from './repository/accessRequests.js';
 export * from './repository/contextDigests.js';
 export * from './repository/memberDiscovery.js';
@@ -60,6 +54,9 @@ export * from './repository/budgetsPrivacy.js';
 export * from './repository/sessions.js';
 export * from './repository/questionDigest.js';
 export * from './repository/knowledgeGaps.js';
+export * from './repository/adminStats.js';
+export * from './repository/contentReports.js';
+export * from './repository/answerFeedback.js';
 
 export interface InteractionInput {
   platform: Platform;
@@ -455,351 +452,6 @@ export async function getInteractionContentByMessageId(
   return rows[0]?.content ?? null;
 }
 
-// --- Super-admin views ---------------------------------------------------------
-
-export async function recentAuditEntries(limit = 20): Promise<
-  Array<{
-    createdAt: Date;
-    platform: string;
-    actorUserId: string;
-    actionKind: string;
-    targetUserId: string | null;
-    success: boolean;
-    result: string | null;
-  }>
-> {
-  const { rows } = await pool.query(
-    `SELECT created_at, platform, actor_user_id, action_kind, target_user_id, success, result
-       FROM admin_audit ORDER BY created_at DESC LIMIT $1`,
-    [limit],
-  );
-  return rows.map((r) => ({
-    createdAt: r.created_at,
-    platform: r.platform,
-    actorUserId: r.actor_user_id,
-    actionKind: r.action_kind,
-    targetUserId: r.target_user_id,
-    success: r.success,
-    result: r.result,
-  }));
-}
-
-/** action_kinds an admin-tier `moderation_history` read may surface — allow-list so a
- * future privileged kind (e.g. another `grant_*`) is excluded by default, not by omission. */
-export const MODERATION_ACTION_KINDS = [
-  'warn_user',
-  'timeout_user',
-  'kick_user',
-  'ban_user',
-  'unban_user',
-  'delete_message',
-  'clear_warnings',
-  'announce',
-  'block_user',
-  'unblock_user',
-] as const;
-
-/**
- * Admin-tier view of moderation actions, scoped to `conversationIds` (null = super
- * admin, unrestricted — same convention as recentQuestionClusters). Mirrors
- * recentAuditEntries but additionally surfaces conversation_id (needed both for the
- * scoping filter and so an admin in multiple channels can attribute an entry) and
- * omits `params` (may carry free-text reasons with member PII beyond the target id).
- *
- * `targetUserId`/`actionKind`, when present, narrow the result further — same
- * one-predicate-append technique as listReports's `status` filter — and can never
- * widen it past the mandatory allow-list/scope predicates above.
- */
-export async function recentModerationEntries(
-  conversationIds: readonly string[] | null,
-  limit = 20,
-  targetUserId?: string,
-  actionKind?: (typeof MODERATION_ACTION_KINDS)[number],
-): Promise<
-  Array<{
-    createdAt: Date;
-    platform: string;
-    actorUserId: string;
-    actionKind: string;
-    targetUserId: string | null;
-    conversationId: string | null;
-    success: boolean;
-    result: string | null;
-  }>
-> {
-  const clampedLimit = Math.min(Math.max(Math.trunc(limit) || 20, 1), 100);
-
-  const params: unknown[] = [[...MODERATION_ACTION_KINDS]];
-  let scope = '';
-  if (conversationIds) {
-    params.push([...conversationIds]);
-    scope = `AND conversation_id = ANY($${params.length})`;
-  }
-  let targetFilter = '';
-  if (targetUserId) {
-    params.push(targetUserId);
-    targetFilter = `AND target_user_id = $${params.length}`;
-  }
-  let actionKindFilter = '';
-  if (actionKind) {
-    params.push(actionKind);
-    actionKindFilter = `AND action_kind = $${params.length}`;
-  }
-  params.push(clampedLimit);
-
-  const { rows } = await pool.query(
-    `SELECT created_at, platform, actor_user_id, action_kind, target_user_id, conversation_id, success, result
-       FROM admin_audit
-      WHERE action_kind = ANY($1)
-        ${scope}
-        ${targetFilter}
-        ${actionKindFilter}
-      ORDER BY created_at DESC
-      LIMIT $${params.length}`,
-    params,
-  );
-  return rows.map((r) => ({
-    createdAt: r.created_at,
-    platform: r.platform,
-    actorUserId: r.actor_user_id,
-    actionKind: r.action_kind,
-    targetUserId: r.target_user_id,
-    conversationId: r.conversation_id,
-    success: r.success,
-    result: r.result,
-  }));
-}
-
-/**
- * Per-actor rollup of `admin_audit` over a trailing window (issue #488), the
- * aggregated complement to `recentAuditEntries`'s flat log — answers "who is
- * actually doing moderation/curation work" instead of requiring a super admin
- * to hand-tally raw log lines. Global/unscoped, same as `recentAuditEntries`
- * (a super admin can already read every row via `audit_view`). Reuses
- * `admin_audit_actor_idx (platform, actor_user_id, created_at DESC)` for the
- * `GROUP BY`. Never selects `params` (may carry free-text reasons) — only
- * counts and timestamps. Days clamp mirrors `usageStats`' own shape.
- */
-export async function adminActivitySummary(days = 30): Promise<
-  Array<{
-    platform: Platform;
-    actorUserId: string;
-    actionCount: number;
-    successCount: number;
-    failureCount: number;
-    lastActionAt: Date;
-  }>
-> {
-  const clampedDays = Math.min(Math.max(Math.trunc(days) || 30, 1), 365);
-  // Exclude the auto-enroll sentinel (issue #606 review): this rollup exists to
-  // rank HUMAN moderation/curation work by volume, but with
-  // DISCORD_AUTO_ENROLL_MEMBERS on, every join writes a system-actor audit row.
-  // On an active server that actor would dominate the top of the report and
-  // bury real admin activity — the opposite of what the tool is for. The rows
-  // stay in admin_audit (still visible via audit_view); they're only kept out
-  // of this ranking.
-  const { rows } = await pool.query(
-    `SELECT platform, actor_user_id,
-            count(*) AS action_count,
-            count(*) FILTER (WHERE success) AS success_count,
-            count(*) FILTER (WHERE NOT success) AS failure_count,
-            max(created_at) AS last_action_at
-       FROM admin_audit
-      WHERE created_at >= now() - $1::interval
-        AND actor_user_id <> $2
-      GROUP BY platform, actor_user_id
-      ORDER BY count(*) DESC`,
-    [`${clampedDays} days`, AUTO_ENROLL_ACTOR],
-  );
-  return rows.map((r) => ({
-    platform: r.platform as Platform,
-    actorUserId: r.actor_user_id as string,
-    actionCount: Number(r.action_count),
-    successCount: Number(r.success_count),
-    failureCount: Number(r.failure_count),
-    lastActionAt: r.last_action_at as Date,
-  }));
-}
-
-export async function usageStats(
-  days = 7,
-  platform?: Platform,
-): Promise<{
-  inbound: number;
-  outbound: number;
-  costUsd: number;
-  byPlatform: Array<{ platform: Platform; inbound: number; outbound: number; costUsd: number }>;
-  topUsers: Array<{ userId: string; userName: string | null; messages: number }>;
-  costByRole: Array<{ role: Tier; costUsd: number; replies: number }>;
-  backgroundCostUsd: number;
-  shortcutHits: { total: number; byKind: Array<{ kind: string; count: number }> };
-  backgroundCostByJob: Array<{ job: string; costUsd: number }>;
-  cacheUsage: { readTokens: number; creationTokens: number };
-  autoAnswerUsage: { count: number; costUsd: number };
-  costByModel: Array<{ model: string; costUsd: number; replies: number }>;
-}> {
-  const clampedDays = Math.min(Math.max(Math.trunc(days) || 7, 1), 365);
-  const interval = `${clampedDays} days`;
-  // Optional platform scoping (issue #647), mirroring `engagementStats`'s
-  // dynamic-placeholder pattern: only `totals`/`top`/`byRole` — the three
-  // fields named by #580's growth path — take the filter. `byPlatform` and
-  // the background/cache/shortcut/auto-answer aggregates below stay
-  // global-only by design (they aren't platform-attributed, or scoping them
-  // is out of scope per the approved proposal).
-  const scopeParams: unknown[] = [interval];
-  let scopeFilter = '';
-  if (platform) {
-    scopeParams.push(platform);
-    scopeFilter = ` AND platform = $${scopeParams.length}`;
-  }
-  const { rows: totals } = await pool.query(
-    `SELECT
-       count(*) FILTER (WHERE direction = 'inbound') AS inbound,
-       count(*) FILTER (WHERE direction = 'outbound') AS outbound,
-       coalesce(sum(cost_usd), 0) AS cost
-     FROM interactions WHERE created_at > now() - $1::interval${scopeFilter}`,
-    scopeParams,
-  );
-  // Per-platform split of the same `totals` query above (issue #580) — same
-  // table/window/direction semantics, differing only by `GROUP BY platform`,
-  // so summed rows equal `totals` by construction. Ordered by volume desc
-  // (then platform name) so the tool's output line is deterministic. Always
-  // computed unfiltered — `formatUsageStats` omits rendering it when a
-  // `platform` filter is active (issue #647), since a single-platform view
-  // makes the breakdown line redundant.
-  const { rows: byPlatform } = await pool.query(
-    `SELECT platform,
-       count(*) FILTER (WHERE direction = 'inbound') AS inbound,
-       count(*) FILTER (WHERE direction = 'outbound') AS outbound,
-       coalesce(sum(cost_usd), 0) AS cost
-     FROM interactions WHERE created_at > now() - $1::interval
-     GROUP BY platform
-     ORDER BY count(*) DESC, platform`,
-    [interval],
-  );
-  const { rows: top } = await pool.query(
-    `SELECT user_id, max(user_name) AS user_name, count(*) AS n
-       FROM interactions
-      WHERE direction = 'inbound' AND created_at > now() - $1::interval${scopeFilter}
-      GROUP BY user_id ORDER BY n DESC LIMIT 5`,
-    scopeParams,
-  );
-  const { rows: byRole } = await pool.query(
-    `SELECT role, coalesce(sum(cost_usd), 0) AS cost, count(*) AS n
-       FROM interactions
-      WHERE direction = 'outbound' AND created_at > now() - $1::interval${scopeFilter}
-      GROUP BY role ORDER BY sum(cost_usd) DESC, role`,
-    scopeParams,
-  );
-  // Cache-hit/-write token telemetry (issue #522): sums the `meta` JSONB
-  // keys `core.ts`/`router.ts` write per outbound row (issue #508's read,
-  // threaded through). Same table/window/direction filter as `byRole` above,
-  // just one more SUM() aggregate over it — same cost class as the existing
-  // `backgroundCostByJob`/`shortcutHits` aggregates it sits beside. Rows that
-  // never got either key (pre-#522, or a turn with no/zero usage) contribute
-  // 0 via `coalesce`, not null.
-  const { rows: cache } = await pool.query(
-    `SELECT
-       coalesce(sum((meta->>'cacheReadTokens')::bigint), 0) AS read_tokens,
-       coalesce(sum((meta->>'cacheCreationTokens')::bigint), 0) AS creation_tokens
-     FROM interactions
-     WHERE direction = 'outbound' AND created_at > now() - $1::interval`,
-    [interval],
-  );
-  // Auto-answer cost visibility (issue #552): mirrors the cache-usage
-  // aggregate immediately above — same table/window/direction filter, one
-  // more pair of SUM()/COUNT() over the `meta->>'autoAnswer'` key
-  // `router.ts`'s `respond()` now stamps. Rows predating this change (or any
-  // non-auto-answer reply) carry no such key and contribute 0 via `coalesce`.
-  const { rows: autoAnswer } = await pool.query(
-    `SELECT
-       coalesce(count(*) FILTER (WHERE meta->>'autoAnswer' = 'true'), 0) AS count,
-       coalesce(sum(cost_usd) FILTER (WHERE meta->>'autoAnswer' = 'true'), 0) AS cost
-     FROM interactions
-     WHERE direction = 'outbound' AND created_at > now() - $1::interval`,
-    [interval],
-  );
-  // Per-model cost telemetry (issue #792): sums the `meta->'modelUsage'`
-  // JSONB object `core.ts`/`router.ts` write per outbound row, keyed by
-  // canonical model id. `jsonb_each_text` is a strict set-returning
-  // function, so a row with no `modelUsage` key (meta->'modelUsage' is SQL
-  // NULL) contributes zero expanded rows rather than needing an explicit
-  // `meta ? 'modelUsage'` filter — same table/window/direction filter as the
-  // cache/auto-answer aggregates above, one more LATERAL aggregate over it.
-  const { rows: costByModelRows } = await pool.query(
-    `SELECT mu.key AS model, coalesce(sum(mu.value::numeric), 0) AS cost, count(*) AS n
-     FROM interactions
-     CROSS JOIN LATERAL jsonb_each_text(meta->'modelUsage') AS mu(key, value)
-     WHERE direction = 'outbound' AND created_at > now() - $1::interval
-     GROUP BY mu.key
-     ORDER BY cost DESC, mu.key`,
-    [interval],
-  );
-  const background = await sumBackgroundJobCosts(clampedDays);
-  const shortcuts = await sumShortcutHits(clampedDays);
-  return {
-    inbound: Number(totals[0].inbound),
-    outbound: Number(totals[0].outbound),
-    costUsd: Number(totals[0].cost),
-    byPlatform: byPlatform.map((r) => ({
-      platform: r.platform as Platform,
-      inbound: Number(r.inbound),
-      outbound: Number(r.outbound),
-      costUsd: Number(r.cost),
-    })),
-    topUsers: top.map((r) => ({ userId: r.user_id, userName: r.user_name, messages: Number(r.n) })),
-    costByRole: byRole.map((r) => ({ role: r.role, costUsd: Number(r.cost), replies: Number(r.n) })),
-    backgroundCostUsd: background.total,
-    shortcutHits: shortcuts,
-    backgroundCostByJob: background.byJob,
-    cacheUsage: {
-      readTokens: Number(cache[0].read_tokens),
-      creationTokens: Number(cache[0].creation_tokens),
-    },
-    autoAnswerUsage: {
-      count: Number(autoAnswer[0].count),
-      costUsd: Number(autoAnswer[0].cost),
-    },
-    costByModel: costByModelRows.map((r) => ({
-      model: r.model as string,
-      costUsd: Number(r.cost),
-      replies: Number(r.n),
-    })),
-  };
-}
-
-// --- Background job costs ---------------------------------------------------
-
-export type BackgroundJob = 'moderation_llm' | 'context_builder' | 'knowledge_refresh';
-
-/**
- * Records the cost of a standalone background `query()` call (issue #401) —
- * one of the three that spend from the shared Max pool but write no
- * `interactions` row, so `usageStats()` would otherwise never see them.
- * Callers are expected to fire this without awaiting and swallow rejections
- * (see `classifyAbuseWithLlm`/`summarizeCluster`/`researchTopic`), matching
- * this codebase's non-blocking-telemetry convention — a failed write must
- * never block or fail the underlying job.
- */
-export async function recordBackgroundJobCost(job: BackgroundJob, costUsd: number): Promise<void> {
-  await pool.query(`INSERT INTO background_job_costs (job, cost_usd) VALUES ($1, $2)`, [job, costUsd]);
-}
-
-export async function sumBackgroundJobCosts(
-  days = 7,
-): Promise<{ total: number; byJob: Array<{ job: string; costUsd: number }> }> {
-  const clampedDays = Math.min(Math.max(Math.trunc(days) || 7, 1), 365);
-  const { rows } = await pool.query(
-    `SELECT job, coalesce(sum(cost_usd), 0) AS cost
-       FROM background_job_costs
-      WHERE created_at > now() - $1::interval
-      GROUP BY job ORDER BY job`,
-    [`${clampedDays} days`],
-  );
-  const byJob = rows.map((r) => ({ job: r.job as string, costUsd: Number(r.cost) }));
-  return { total: byJob.reduce((sum, r) => sum + r.costUsd, 0), byJob };
-}
-
 // --- Member-facing weekly digest freshness guard (issue #645) --------------
 
 /**
@@ -915,1175 +567,4 @@ export async function listReleaseWatchUpdatesSince(
     if (!byPage.has(page)) byPage.set(page, { pageTitle: page, sourceUrl: row.source_url });
   }
   return [...byPage.values()].slice(0, clampedLimit);
-}
-
-// --- Content reports (member-facing abuse/spam intake) -----------------------
-
-/** Per-reporter cap on new reports within a rolling window (anti-griefing on the admin queue). */
-export const REPORT_RATE_LIMIT_PER_DAY = 5;
-
-export type ContentReportStatus = 'open' | 'resolved' | 'dismissed' | 'withdrawn';
-
-export interface ContentReport {
-  id: number;
-  platform: Platform;
-  reporterUserId: string;
-  reporterName: string | null;
-  conversationId: string;
-  targetUserId: string | null;
-  messageId: string | null;
-  reason: string;
-  status: ContentReportStatus;
-  createdAt: Date;
-  resolvedBy: string | null;
-  resolvedAt: Date | null;
-}
-
-function mapContentReport(r: {
-  id: number | string;
-  platform: string;
-  reporter_user_id: string;
-  reporter_name: string | null;
-  conversation_id: string;
-  target_user_id: string | null;
-  message_id: string | null;
-  reason: string;
-  status: string;
-  created_at: Date;
-  resolved_by: string | null;
-  resolved_at: Date | null;
-}): ContentReport {
-  return {
-    id: Number(r.id),
-    platform: r.platform as Platform,
-    reporterUserId: r.reporter_user_id,
-    reporterName: r.reporter_name,
-    conversationId: r.conversation_id,
-    targetUserId: r.target_user_id,
-    messageId: r.message_id,
-    reason: r.reason,
-    status: r.status as ContentReportStatus,
-    createdAt: r.created_at,
-    resolvedBy: r.resolved_by,
-    resolvedAt: r.resolved_at,
-  };
-}
-
-/**
- * Record a member's report, enforcing a DB-backed rolling-24h cap per
- * reporter (COUNT(*) over content_reports, not an in-memory counter — the
- * only existing rate limiter, router.ts's per-message map, resets on
- * restart and would let a bounce bypass the cap). Returns null when the
- * caller is at/over the cap; the tool layer turns that into a polite refusal.
- */
-export async function createContentReport(input: {
-  platform: Platform;
-  reporterUserId: string;
-  reporterName?: string;
-  conversationId: string;
-  targetUserId?: string;
-  messageId?: string;
-  reason: string;
-  /** Filed from a 1:1 DM? Defaults to false (matching the column default) for callers that don't track it. */
-  isDirect?: boolean;
-}): Promise<{ id: number } | null> {
-  const { rows } = await pool.query(
-    `WITH recent AS (
-       SELECT count(*) AS n FROM content_reports
-        WHERE platform = $1 AND reporter_user_id = $2
-          AND created_at > now() - interval '24 hours'
-     )
-     INSERT INTO content_reports
-       (platform, reporter_user_id, reporter_name, conversation_id, target_user_id, message_id, reason, is_dm)
-     SELECT $1, $2, $3, $4, $5, $6, $7, $9
-      WHERE (SELECT n FROM recent) < $8
-     RETURNING id`,
-    [
-      input.platform,
-      input.reporterUserId,
-      input.reporterName ?? null,
-      input.conversationId,
-      input.targetUserId ?? null,
-      input.messageId ?? null,
-      input.reason.slice(0, 500),
-      REPORT_RATE_LIMIT_PER_DAY,
-      input.isDirect ?? false,
-    ],
-  );
-  return rows[0] ? { id: Number(rows[0].id) } : null;
-}
-
-/**
- * Count of DM reports the given reporter has filed naming the given target
- * within the last `windowDays` — narrows the SECURITY.md-documented residual
- * risk that a member who knows an admin's platform id (e.g. from an
- * @-mention) can repeatedly name them in unrelated DM reports, quietly
- * blinding that admin via the accused-admin exclusion in `listReports`/
- * `countOpenReports`/`resolveContentReport` (issue #197), with nothing
- * surfacing the pattern (issue #305). Scoped exactly to `(platform,
- * reporter_user_id, target_user_id, is_dm = true)` — a different platform,
- * reporter, target, or a non-DM report never contributes. Served by the
- * existing `content_reports_reporter_rate_idx (platform, reporter_user_id,
- * created_at DESC)` for its `(platform, reporter_user_id)` prefix; report
- * volume is already capped at `REPORT_RATE_LIMIT_PER_DAY` per reporter per
- * rolling 24h, so this stays cheap regardless of call frequency.
- */
-export async function countRecentDmReportsByReporterAndTarget(
-  platform: Platform,
-  reporterUserId: string,
-  targetUserId: string,
-  windowDays = 30,
-): Promise<number> {
-  const { rows } = await pool.query(
-    `SELECT count(*) AS n FROM content_reports
-      WHERE platform = $1 AND reporter_user_id = $2 AND target_user_id = $3
-        AND is_dm = true AND created_at > now() - ($4 || ' days')::interval`,
-    [platform, reporterUserId, targetUserId, String(windowDays)],
-  );
-  return Number(rows[0].n);
-}
-
-/**
- * Admin-tier view of reports, scoped to `conversationIds` (null = super
- * admin, unrestricted — same convention as recentModerationEntries). A
- * report filed from a 1:1 DM (`is_dm`) has no conversation any ordinary
- * admin can naturally be scoped to (each DM is unique per member), so it is
- * additionally surfaced to a scoped admin via `OR is_dm` — except one filed
- * against that very admin (`target_user_id`), which stays reachable only by
- * a super admin so an accused admin can't see or dismiss a report about
- * themselves (issue #197). `viewerUserIds` is the calling admin's own id
- * PLUS every identity linked to them via `link_member` — a single raw id
- * would let a dual-presence admin (Discord + WhatsApp, exactly the case
- * `link_member` exists for) see a DM report filed against their *other*
- * platform identity, since that id `IS DISTINCT FROM` their current-platform
- * id. Omitting it leaves DM-originated reports invisible to a scoped admin,
- * same as before #197 — never widen scope without it.
- *
- * `targetUserId`, when present, narrows the result further — same
- * one-predicate-append technique as recentModerationEntries's `targetUserId`
- * filter — and is appended AFTER the accused-admin exclusion above, so it can
- * only intersect an already-scoped result set, never widen it (issue #463).
- */
-export async function listReports(
-  conversationIds: readonly string[] | null,
-  status?: ContentReportStatus,
-  limit = 50,
-  viewerUserIds?: readonly string[],
-  targetUserId?: string,
-): Promise<ContentReport[]> {
-  const params: unknown[] = [];
-  const filters: string[] = [];
-  if (conversationIds) {
-    params.push([...conversationIds]);
-    const convoIdx = params.length;
-    if (viewerUserIds && viewerUserIds.length > 0) {
-      params.push([...viewerUserIds]);
-      filters.push(
-        `(conversation_id = ANY($${convoIdx}) OR (is_dm AND (target_user_id IS NULL OR target_user_id <> ALL($${params.length}))))`,
-      );
-    } else {
-      filters.push(`conversation_id = ANY($${convoIdx})`);
-    }
-  }
-  if (status) {
-    params.push(status);
-    filters.push(`status = $${params.length}`);
-  }
-  if (targetUserId) {
-    params.push(targetUserId);
-    filters.push(`target_user_id = $${params.length}`);
-  }
-  const clampedLimit = Math.min(Math.max(Math.trunc(limit) || 50, 1), 200);
-  params.push(clampedLimit);
-  const where = filters.length ? `WHERE ${filters.join(' AND ')}` : '';
-
-  const { rows } = await pool.query(
-    `SELECT id, platform, reporter_user_id, reporter_name, conversation_id, target_user_id,
-            message_id, reason, status, created_at, resolved_by, resolved_at
-       FROM content_reports
-       ${where}
-      ORDER BY created_at DESC
-      LIMIT $${params.length}`,
-    params,
-  );
-  return rows.map(mapContentReport);
-}
-
-/**
- * Exact open-report count, scoped like `listReports` (`conversationIds`
- * null = unrestricted/super admin, `viewerUserIds` drives the same
- * linked-identity-aware accused-admin exclusion on the `OR is_dm` broadening
- * — see `listReports`) — a dedicated `COUNT(*)` rather than `(await
- * listReports(scope, 'open')).length`, which would silently understate a
- * backlog past that function's clamped (≤200) `limit`.
- */
-export async function countOpenReports(
-  conversationIds: readonly string[] | null,
-  viewerUserIds?: readonly string[],
-): Promise<number> {
-  const params: unknown[] = [];
-  const filters: string[] = [`status = 'open'`];
-  if (conversationIds) {
-    params.push([...conversationIds]);
-    const convoIdx = params.length;
-    if (viewerUserIds && viewerUserIds.length > 0) {
-      params.push([...viewerUserIds]);
-      filters.push(
-        `(conversation_id = ANY($${convoIdx}) OR (is_dm AND (target_user_id IS NULL OR target_user_id <> ALL($${params.length}))))`,
-      );
-    } else {
-      filters.push(`conversation_id = ANY($${convoIdx})`);
-    }
-  }
-  const { rows } = await pool.query(
-    `SELECT count(*) AS n FROM content_reports WHERE ${filters.join(' AND ')}`,
-    params,
-  );
-  return Number(rows[0].n);
-}
-
-/**
- * Whole-day age of the oldest still-open report visible to this admin — the
- * same `MIN(created_at)` oldest-age mechanic `oldestAccessRequestAgeDays`
- * (#515) applies to access requests, over exactly the scoped row set
- * `countOpenReports` counts (issue #450). Builds the identical
- * `status = 'open'` + conversation/DM scoping predicate `countOpenReports`
- * does — so a report filed in a conversation this admin doesn't participate
- * in (or a DM report against the admin themselves) can never influence the
- * age they see, same as it can't influence their count. `MIN` over an empty
- * scoped set is `null`, never `0`, and is returned as-is so a digest reader
- * can never mistake "no open reports" for "one that just arrived".
- */
-export async function oldestOpenReportAgeDays(
-  conversationIds: readonly string[] | null,
-  viewerUserIds?: readonly string[],
-): Promise<number | null> {
-  const params: unknown[] = [];
-  const filters: string[] = [`status = 'open'`];
-  if (conversationIds) {
-    params.push([...conversationIds]);
-    const convoIdx = params.length;
-    if (viewerUserIds && viewerUserIds.length > 0) {
-      params.push([...viewerUserIds]);
-      filters.push(
-        `(conversation_id = ANY($${convoIdx}) OR (is_dm AND (target_user_id IS NULL OR target_user_id <> ALL($${params.length}))))`,
-      );
-    } else {
-      filters.push(`conversation_id = ANY($${convoIdx})`);
-    }
-  }
-  const { rows } = await pool.query(
-    `SELECT EXTRACT(DAY FROM now() - MIN(created_at))::int AS age_days FROM content_reports WHERE ${filters.join(' AND ')}`,
-    params,
-  );
-  const ageDays = rows[0]?.age_days;
-  return ageDays === null || ageDays === undefined ? null : Number(ageDays);
-}
-
-/**
- * Count knowledge-search gaps (#208) recorded in the given conversations within
- * the last `days`, for the weekly admin digest (#246). **Conversation-scoped**
- * — unlike the guild-wide stale/access/suggestion counts — because
- * `knowledge_gaps` has a `conversation_id`, so an admin never sees gap volume
- * from a conversation they don't participate in (mirrors `countOpenReports`'s
- * scoping). A true `COUNT(*)`, never `.length` of a `LIMIT`-bounded list, so a
- * backlog larger than `list_knowledge_gaps`' own limit is not understated.
- * Excludes `resolved_at IS NOT NULL` rows (issue #422), same as
- * `recentKnowledgeGapClusters` — a resolved gap drops out of the digest
- * count immediately.
- */
-export async function countKnowledgeGaps(conversationIds: readonly string[], days: number): Promise<number> {
-  if (conversationIds.length === 0) return 0;
-  const { rows } = await pool.query(
-    `SELECT count(*) AS n FROM knowledge_gaps
-      WHERE conversation_id = ANY($1)
-        AND resolved_at IS NULL
-        AND created_at >= now() - ($2 || ' days')::interval`,
-    [[...conversationIds], String(days)],
-  );
-  return Number(rows[0].n);
-}
-
-/**
- * Count ESCALATED knowledge gaps (issue #514) recorded in the given
- * conversations within the last `days`, for the weekly admin digest — the
- * subset of `countKnowledgeGaps` written by `recordEscalatedKnowledgeGap`
- * (a confirmed, member-initiated escalation rather than a passive miss).
- * Mirrors `countKnowledgeGaps` exactly (conversation-scoped, day-windowed,
- * `resolved_at IS NULL`, a true `COUNT(*)`) plus `AND escalated = true`.
- */
-export async function countEscalatedKnowledgeGaps(
-  conversationIds: readonly string[],
-  days: number,
-): Promise<number> {
-  if (conversationIds.length === 0) return 0;
-  const { rows } = await pool.query(
-    `SELECT count(*) AS n FROM knowledge_gaps
-      WHERE conversation_id = ANY($1)
-        AND resolved_at IS NULL
-        AND escalated = true
-        AND created_at >= now() - ($2 || ' days')::interval`,
-    [[...conversationIds], String(days)],
-  );
-  return Number(rows[0].n);
-}
-
-/**
- * Count outbound replies in `conversationIds` over the last `days` that hit
- * `AGENT_MAX_TURNS`/`AGENT_MAX_TURNS_MEMBER` before finishing, for the weekly
- * admin digest (#371). Counts both the primary `maxTurnsExceeded: true` stamp
- * (router.ts's outbound-record call) and the `repeatMaxTurnsShortcut: true`
- * stamp (#306) — each is a distinct member-facing wall-hit. **Conversation-
- * scoped** and a true `COUNT(*)`, mirroring `countKnowledgeGaps` exactly.
- */
-export async function countMaxTurnsFailures(
-  conversationIds: readonly string[],
-  days: number,
-): Promise<number> {
-  if (conversationIds.length === 0) return 0;
-  const { rows } = await pool.query(
-    `SELECT count(*) AS n FROM interactions
-      WHERE direction = 'outbound'
-        AND conversation_id = ANY($1)
-        AND created_at >= now() - ($2 || ' days')::interval
-        AND (meta->>'maxTurnsExceeded' = 'true' OR meta->>'repeatMaxTurnsShortcut' = 'true')`,
-    [[...conversationIds], String(days)],
-  );
-  return Number(rows[0].n);
-}
-
-/**
- * Count unhelpful ratings on GENERAL-KNOWLEDGE answers (issue #563) — the
- * `meta->>'knowledgeEntryId' IS NULL` complement `countLowRatedKnowledge`/
- * `listKnowledgeFeedbackSummary` deliberately exclude (their own doc
- * comments: "Ratings on interactions with no `knowledgeEntryId` are still
- * excluded"). A general-knowledge answer has no community-curated grounding
- * to re-check, unlike a KB-attributed one — the highest accuracy-risk bucket
- * per VISION's answer-quality theme — so this is the missing push signal for
- * it. Modelled on `countMaxTurnsFailures`'s rolling-window, conversation-
- * scoped, true-`COUNT(*)` shape rather than `countLowRatedKnowledge`'s
- * per-entity backlog shape: free-text general-knowledge answers have no
- * stable grouping key to bucket repeated ratings against.
- *
- * The JOIN to `interactions` (rather than a `meta` subquery) means a row
- * whose `interaction_id` is NULL — e.g. after the rated reply was purged via
- * `forget_me`/`purge_user_data`, which sets `answer_feedback.interaction_id`
- * to NULL on delete (schema.sql) — is excluded: with no interaction left to
- * join, there's nothing to classify as grounded or ungrounded.
- */
-export async function countGeneralUnhelpfulAnswers(
-  conversationIds: readonly string[],
-  days: number,
-): Promise<number> {
-  if (conversationIds.length === 0) return 0;
-  const { rows } = await pool.query(
-    `SELECT count(*) AS n
-       FROM answer_feedback
-       JOIN interactions ON interactions.id = answer_feedback.interaction_id
-      WHERE answer_feedback.helpful = false
-        AND answer_feedback.conversation_id = ANY($1)
-        AND answer_feedback.created_at >= now() - ($2 || ' days')::interval
-        AND (interactions.meta->>'knowledgeEntryId') IS NULL`,
-    [[...conversationIds], String(days)],
-  );
-  return Number(rows[0].n);
-}
-
-export interface AnswerFeedbackWeeklySummary {
-  helpful: number;
-  total: number;
-}
-
-/**
- * Overall this-week helpful-rate across EVERY rated answer (issue #653) —
- * VISION's own named answer-quality north star, currently invisible: neither
- * `countGeneralUnhelpfulAnswers` (#563, ungrounded-only unhelpful COUNT) nor
- * `answerFeedbackOriginSummary` (#592, cumulative origin split) answers "what
- * fraction of all rated answers this week were helpful?" — a knowledge-
- * grounded answer rated unhelpful lands in neither one's numerator or
- * denominator. This read is deliberately **unfiltered** by
- * `knowledgeEntryId`/origin/`autoAnswer` — every rated answer this week,
- * counted once — the distinct, all-answer-types denominator neither sibling
- * signal covers.
- *
- * Same rolling-window, conversation-scoped, true-`COUNT(*)` shape as
- * `countGeneralUnhelpfulAnswers`/`countMaxTurnsFailures`, including the same
- * JOIN-to-`interactions` exclusion of a purged rating (`interaction_id` set
- * NULL by `forget_me`/`purge_user_data`) and the same empty-`conversationIds`
- * early return (zero-counts, no query issued).
- */
-export async function answerFeedbackWeeklySummary(
-  conversationIds: readonly string[],
-  days: number,
-): Promise<AnswerFeedbackWeeklySummary> {
-  if (conversationIds.length === 0) return { helpful: 0, total: 0 };
-  const { rows } = await pool.query(
-    `SELECT
-       count(*) FILTER (WHERE answer_feedback.helpful) AS helpful,
-       count(*) AS total
-       FROM answer_feedback
-       JOIN interactions ON interactions.id = answer_feedback.interaction_id
-      WHERE answer_feedback.conversation_id = ANY($1)
-        AND answer_feedback.created_at >= now() - ($2 || ' days')::interval`,
-    [[...conversationIds], String(days)],
-  );
-  return { helpful: Number(rows[0].helpful), total: Number(rows[0].total) };
-}
-
-export interface UnhelpfulFeedbackCluster {
-  representative: string;
-  count: number;
-}
-
-/**
- * Greedily cluster recent unhelpful `answer_feedback` comments by embedding
- * similarity — the second, still-missing half of VISION's answer-quality
- * north star ("thumbs-down themes shrinking in the digests", issue #724).
- * Mirrors `recentQuestionClusters`/`recentKnowledgeGapClusters` exactly (same
- * greedy clustering code, same `QUESTION_CLUSTER_SIMILARITY_THRESHOLD`, same
- * `count >= 2` theme floor, same conversation-scoping convention — null =
- * super admin, unrestricted), but sourced from `answer_feedback` instead of
- * `interactions`/`knowledge_gaps`.
- *
- * Deliberately NOT filtered by `knowledgeEntryId` — both grounded and
- * ungrounded unhelpful answers are included, closing the exact gap
- * `listKnowledgeFeedbackSummary` (grounded-only) and
- * `countGeneralUnhelpfulAnswers` (ungrounded-only, no comment text) each
- * leave open on their own.
- *
- * `answer_feedback` has no persisted `embedding` column (issue #706's
- * call-time-embed precedent, not a schema change): each qualifying comment is
- * embedded here, at read time, via the same local/offline `embed()`. Volume
- * is already double-bounded by `RATE_ANSWER_DAILY_LIMIT` and the
- * `helpful = false AND comment IS NOT NULL` filter, so this is realistically
- * a handful of rows per admin scope per digest window — never a bulk
- * backfill. A row whose `embed()` call fails is skipped (logged, not
- * thrown) exactly like `recordKnowledgeGap`'s embed-failure handling, so a
- * transient embedding-model outage degrades to fewer clusters, never a
- * crash.
- */
-export async function recentUnhelpfulFeedbackClusters(
-  conversationIds: readonly string[] | null,
-  days = 7,
-  limit = 10,
-): Promise<UnhelpfulFeedbackCluster[]> {
-  const clampedDays = Math.min(Math.max(Math.trunc(days) || 7, 1), 30);
-  const clampedLimit = Math.min(Math.max(Math.trunc(limit) || 10, 1), 50);
-
-  const params: unknown[] = [`${clampedDays} days`];
-  let scope = '';
-  if (conversationIds) {
-    params.push([...conversationIds]);
-    scope = `AND conversation_id = ANY($${params.length})`;
-  }
-
-  const { rows } = await pool.query(
-    `SELECT comment
-       FROM answer_feedback
-      WHERE helpful = false
-        AND comment IS NOT NULL
-        AND created_at > now() - $1::interval
-        ${scope}
-      ORDER BY created_at ASC`,
-    params,
-  );
-
-  const clusters: Array<{ representative: string; embedding: number[]; count: number }> = [];
-  for (const row of rows) {
-    const comment = row.comment as string;
-    let vec: number[] | null = null;
-    try {
-      vec = await embed(comment);
-    } catch (err) {
-      logger.warn({ err }, 'Embedding failed for unhelpful-feedback clustering');
-    }
-    if (!vec) continue;
-    const match = clusters.find((c) => cosineSim(c.embedding, vec) >= QUESTION_CLUSTER_SIMILARITY_THRESHOLD);
-    if (match) {
-      match.count += 1;
-    } else {
-      clusters.push({ representative: comment, embedding: vec, count: 1 });
-    }
-  }
-
-  return clusters
-    .filter((c) => c.count >= 2)
-    .sort((a, b) => b.count - a.count)
-    .slice(0, clampedLimit)
-    .map((c) => ({ representative: c.representative, count: c.count }));
-}
-
-/**
- * Flip a report's status (resolve/dismiss) — non-destructive, no CONFIRM
- * needed (mirrors warn_user's low-blast-radius treatment). Optionally scoped
- * to `conversationIds` so an admin can only resolve reports from
- * conversations they actually participate in (same invariant as `moderate`/
- * `announce`) — broadened by `OR is_dm` for the same reason as `listReports`
- * (a DM report has no conversation an ordinary admin is ever scoped to), with
- * the same accused-admin exclusion: the acting admin (and every identity
- * linked to them via `link_member`, passed as `viewerUserIds`) can never
- * resolve a DM report filed against itself — that stays super-admin-only, so
- * an accused admin can't dismiss a report about themselves, and can't slip
- * past the exclusion by resolving from a linked other-platform identity
- * (issue #197). `resolvedBy` still records the single acting id.
- * `viewerUserIds` defaults to `[resolvedBy]` when omitted. Returns the
- * resolved row's platform/reporterUserId/reason (so the caller can notify the
- * reporter, issue #120 — same "RETURNING" shape as `resolveSuggestion`) or
- * null if no matching row was found (unknown id, or the id exists but is
- * outside the caller's scope) — same "no match" signal the old boolean return
- * gave.
- */
-export async function resolveContentReport(
-  id: number,
-  status: 'resolved' | 'dismissed',
-  resolvedBy: string,
-  conversationIds?: readonly string[],
-  viewerUserIds?: readonly string[],
-): Promise<{ platform: Platform; reporterUserId: string; reason: string } | null> {
-  const params: unknown[] = [id, status, resolvedBy];
-  let scope = '';
-  if (conversationIds) {
-    params.push([...conversationIds]);
-    const convoIdx = params.length;
-    params.push([...(viewerUserIds && viewerUserIds.length > 0 ? viewerUserIds : [resolvedBy])]);
-    scope = `AND (conversation_id = ANY($${convoIdx}) OR (is_dm AND (target_user_id IS NULL OR target_user_id <> ALL($${params.length}))))`;
-  }
-  const { rows } = await pool.query(
-    `UPDATE content_reports
-        SET status = $2, resolved_by = $3, resolved_at = now()
-      WHERE id = $1 ${scope}
-      RETURNING platform, reporter_user_id, reason`,
-    params,
-  );
-  return rows[0]
-    ? {
-        platform: rows[0].platform as Platform,
-        reporterUserId: rows[0].reporter_user_id,
-        reason: rows[0].reason,
-      }
-    : null;
-}
-
-/**
- * Let a reporter withdraw their OWN still-open report(s). Members had no
- * self-service way to retract a report (e.g. one filed as a joke) — they had
- * to ask an admin, who then dismisses it, which is awkward when the report is
- * *about* an admin. Strictly scoped to the caller's own identity: the
- * `reporter_user_id = $2` predicate means a member can only ever touch reports
- * they themselves filed, never anyone else's. Non-destructive — the row is
- * marked `'withdrawn'` (distinct from an admin-initiated `'dismissed'`) and
- * kept on record for accountability, never deleted. `resolved_by` is set to
- * the reporter's own id (they did it). Returns the withdrawn ids (empty array
- * if the caller had no open reports).
- */
-export async function withdrawOwnReports(platform: Platform, reporterUserId: string): Promise<number[]> {
-  const { rows } = await pool.query(
-    `UPDATE content_reports
-        SET status = 'withdrawn', resolved_by = $2, resolved_at = now()
-      WHERE platform = $1 AND reporter_user_id = $2 AND status = 'open'
-      RETURNING id`,
-    [platform, reporterUserId],
-  );
-  return rows.map((r) => Number(r.id));
-}
-
-/**
- * Self-scoped read of a member's OWN content reports — mirrors
- * listOwnSuggestions above and reuses withdrawOwnReports's
- * `reporter_user_id = $2` scoping, so a member can only ever see reports they
- * themselves filed, never anyone else's.
- */
-export async function listOwnReports(
-  platform: Platform,
-  reporterUserId: string,
-  limit = 10,
-): Promise<ContentReport[]> {
-  const clampedLimit = Math.min(Math.max(Math.trunc(limit) || 10, 1), 50);
-  const { rows } = await pool.query(
-    `SELECT id, platform, reporter_user_id, reporter_name, conversation_id, target_user_id,
-            message_id, reason, status, created_at, resolved_by, resolved_at
-       FROM content_reports
-      WHERE platform = $1 AND reporter_user_id = $2
-      ORDER BY created_at DESC
-      LIMIT $3`,
-    [platform, reporterUserId, clampedLimit],
-  );
-  return rows.map(mapContentReport);
-}
-
-// --- Answer feedback (member rating of the bot's own answers, issue #118) ---
-
-/** Per-rater cap on new ratings within a rolling 24h window (DB-backed, same pattern as reports/suggestions). */
-export const RATE_ANSWER_DAILY_LIMIT = 20;
-
-export interface AnswerFeedback {
-  id: number;
-  platform: Platform;
-  conversationId: string;
-  userId: string;
-  interactionId: number | null;
-  helpful: boolean;
-  createdAt: Date;
-  /** The rated answer's text, or `null` when the interaction was since purged. */
-  content: string | null;
-  /** Knowledge entry id the answer was served from, when sent via the deterministic knowledge shortcut. */
-  knowledgeEntryId: number | null;
-  /** Optional free-text reason the rater gave alongside the boolean (issue #354), or `null` if none. */
-  comment: string | null;
-}
-
-function mapAnswerFeedback(r: {
-  id: number | string;
-  platform: string;
-  conversation_id: string;
-  user_id: string;
-  interaction_id: number | string | null;
-  helpful: boolean;
-  created_at: Date;
-  content: string | null;
-  knowledge_entry_id: number | string | null;
-  comment: string | null;
-}): AnswerFeedback {
-  return {
-    id: Number(r.id),
-    platform: r.platform as Platform,
-    conversationId: r.conversation_id,
-    userId: r.user_id,
-    interactionId: r.interaction_id != null ? Number(r.interaction_id) : null,
-    helpful: r.helpful,
-    createdAt: r.created_at,
-    content: r.content,
-    knowledgeEntryId: r.knowledge_entry_id != null ? Number(r.knowledge_entry_id) : null,
-    comment: r.comment,
-  };
-}
-
-/** Max stored length of a `rate_answer` comment (issue #354) — matches the tool's `z.string().max(200)`. */
-export const ANSWER_FEEDBACK_COMMENT_MAX_CHARS = 200;
-
-/**
- * Normalize a `rate_answer` comment before it reaches storage: strip C0
- * control characters (including bare `\r`/`\n`, which `untrusted()` also
- * neutralizes at render time — this is defense in depth, not a substitute)
- * and DEL, trim, then cap length. An empty/whitespace-only or omitted
- * comment stores SQL NULL rather than an empty string.
- */
-function normalizeAnswerFeedbackComment(comment?: string): string | null {
-  if (!comment) return null;
-  // eslint-disable-next-line no-control-regex
-  const stripped = comment.replace(/[\x00-\x1F\x7F]/g, '').trim();
-  if (!stripped) return null;
-  return stripped.slice(0, ANSWER_FEEDBACK_COMMENT_MAX_CHARS);
-}
-
-/**
- * Resolve the interaction a `rate_answer` call should bind to. Prefers the
- * caller's OWN most-recent outbound reply in this conversation
- * (`meta->>'replyToUserId' = userId`, stamped by router.ts on every send),
- * falling back to the conversation's most-recent outbound reply only when no
- * caller-scoped match exists (e.g. a row that predates that meta field).
- * Without the caller-scoped preference, a busy multi-member channel could
- * bind member A's "thanks, that helped" to the answer the bot just gave
- * member B — silently corrupting the signal this feature exists to produce.
- */
-async function resolveAnswerFeedbackTarget(
-  platform: Platform,
-  conversationId: string,
-  userId: string,
-): Promise<number | null> {
-  const { rows } = await pool.query(
-    `SELECT COALESCE(
-       (SELECT id FROM interactions
-         WHERE platform = $1 AND conversation_id = $2 AND direction = 'outbound'
-           AND meta->>'replyToUserId' = $3
-         ORDER BY created_at DESC LIMIT 1),
-       (SELECT id FROM interactions
-         WHERE platform = $1 AND conversation_id = $2 AND direction = 'outbound'
-         ORDER BY created_at DESC LIMIT 1)
-     ) AS id`,
-    [platform, conversationId, userId],
-  );
-  return rows[0]?.id != null ? Number(rows[0].id) : null;
-}
-
-/**
- * Record a member's helpful/unhelpful rating of the bot's most recent answer
- * to them in this conversation. Enforces a DB-backed rolling-24h cap per
- * rater (`RATE_ANSWER_DAILY_LIMIT`), the same restart-proof
- * COUNT(*)-inside-the-insert pattern as createSuggestion/createContentReport
- * (never an in-memory counter). A second rating from the same rater on the
- * same interaction (no new bot reply in between) is not a second vote: the
- * `ON CONFLICT` on `answer_feedback_interaction_rater_idx` (issue #619)
- * updates the existing row's helpful/comment/created_at in place instead of
- * inserting a duplicate — preserving "member changed their mind" while
- * keeping every downstream count (usage_stats, the weekly digest, and the
- * >= 2-rater low-rated-caveat floor) to at most one vote per (rater,
- * answer). The `WHERE interaction_id IS NOT NULL` predicate on the conflict
- * target must match the partial index exactly, or Postgres can't infer it.
- * Returns:
- *  - `{ id }` on success (insert OR update-in-place)
- *  - `'no_recent_answer'` when there is no outbound interaction to bind to
- *    yet (e.g. the member has not been answered in this conversation)
- *  - `'rate_limited'` when the caller is at/over the cap
- */
-export async function createAnswerFeedback(input: {
-  platform: Platform;
-  conversationId: string;
-  userId: string;
-  helpful: boolean;
-  /** Optional free-text reason (issue #354); normalized (control-char-stripped, ≤200 chars) before storage. */
-  comment?: string;
-}): Promise<{ id: number; interactionId: number } | 'no_recent_answer' | 'rate_limited'> {
-  const interactionId = await resolveAnswerFeedbackTarget(input.platform, input.conversationId, input.userId);
-  if (interactionId === null) return 'no_recent_answer';
-
-  const { rows } = await pool.query(
-    `WITH recent AS (
-       SELECT count(*) AS n FROM answer_feedback
-        WHERE platform = $1 AND user_id = $2
-          AND created_at > now() - interval '24 hours'
-     )
-     INSERT INTO answer_feedback (platform, conversation_id, user_id, interaction_id, helpful, comment)
-     SELECT $1, $3, $2, $4, $5, $7
-      WHERE (SELECT n FROM recent) < $6
-     ON CONFLICT (interaction_id, user_id) WHERE interaction_id IS NOT NULL
-     DO UPDATE SET helpful = EXCLUDED.helpful, comment = EXCLUDED.comment, created_at = now()
-     RETURNING id`,
-    [
-      input.platform,
-      input.userId,
-      input.conversationId,
-      interactionId,
-      input.helpful,
-      RATE_ANSWER_DAILY_LIMIT,
-      normalizeAnswerFeedbackComment(input.comment),
-    ],
-  );
-  return rows[0] ? { id: Number(rows[0].id), interactionId } : 'rate_limited';
-}
-
-/**
- * Grounding lookup for the answered-question -> knowledge-base loop (issue
- * #726): given the `interactions.id` of a rated OUTBOUND reply, recovers
- * whether it was knowledge-grounded and, when not, the preceding question it
- * answered. Two reads, no new index: the outbound row's own `content`/
- * `meta->>'knowledgeEntryId'`/`meta->>'replyToUserId'` (the caller `rate_answer`
- * itself bound to via `resolveAnswerFeedbackTarget` above), then — only when a
- * `replyToUserId` exists — the most recent INBOUND row from that SAME member
- * in the SAME conversation at or before the reply's `created_at`, served by
- * `interactions_user_idx (platform, user_id, created_at DESC)`.
- *
- * `questionUserId` deliberately names the addressed member (`replyToUserId`),
- * NOT the `rate_answer` caller — `resolveAnswerFeedbackTarget` can bind a
- * rating to a reply addressed to someone else (the "rater observed, didn't
- * ask" case), so attribution must track the row this returns, never the
- * caller blindly (SECURITY, issue #726 AC10). Returns `null` when
- * `interactionId` doesn't name an outbound row at all; `questionContent`/
- * `questionUserId` are independently `null` (fail closed for the caller) when
- * there is no `replyToUserId` or no qualifying preceding inbound row.
- */
-export async function answerFeedbackGrounding(interactionId: number): Promise<{
-  knowledgeEntryId: number | null;
-  answerContent: string;
-  questionContent: string | null;
-  questionUserId: string | null;
-} | null> {
-  const { rows: outboundRows } = await pool.query(
-    `SELECT platform, conversation_id, content, created_at,
-            (meta->>'knowledgeEntryId')::bigint AS knowledge_entry_id,
-            meta->>'replyToUserId' AS reply_to_user_id
-       FROM interactions
-      WHERE id = $1 AND direction = 'outbound'`,
-    [interactionId],
-  );
-  const outbound = outboundRows[0];
-  if (!outbound) return null;
-
-  let questionContent: string | null = null;
-  let questionUserId: string | null = null;
-  if (outbound.reply_to_user_id) {
-    const { rows: inboundRows } = await pool.query(
-      `SELECT content, user_id FROM interactions
-        WHERE platform = $1 AND conversation_id = $2 AND direction = 'inbound'
-          AND user_id = $3 AND created_at <= $4
-        ORDER BY created_at DESC LIMIT 1`,
-      [outbound.platform, outbound.conversation_id, outbound.reply_to_user_id, outbound.created_at],
-    );
-    if (inboundRows[0]) {
-      questionContent = inboundRows[0].content;
-      questionUserId = inboundRows[0].user_id;
-    }
-  }
-
-  return {
-    knowledgeEntryId: outbound.knowledge_entry_id != null ? Number(outbound.knowledge_entry_id) : null,
-    answerContent: outbound.content,
-    questionContent,
-    questionUserId,
-  };
-}
-
-/**
- * SECURITY (issue #726 follow-up): rater-scoped counterpart to
- * `createKnowledgeTip`'s per-question-author cap. That cap alone bounds how
- * many candidates a single VICTIM's quota can absorb, but not how many
- * DIFFERENT victims one rater can draft against — `resolveAnswerFeedbackTarget`
- * can bind a rating to a reply addressed to someone else (the "rater
- * observed, didn't ask" fallback, see `answerFeedbackGrounding` above and its
- * AC10 test), and `rate_answer`'s own daily cap (`RATE_ANSWER_DAILY_LIMIT`,
- * 20/day) is far looser than any one victim's 3/day quota. Without this, a
- * rater who has never personally been answered could silently drain several
- * other members' entire daily `suggest_knowledge` quota in one busy channel.
- *
- * Counts this rater's OWN `helpful: true` `answer_feedback` rows in the last
- * 24h whose bound interaction was addressed to someone else — i.e., every
- * attempt (successful or not) at the mismatched-attribution drafting path.
- * A matched self-rating (rater === addressed member) is deliberately
- * excluded: that case is already bounded by `createKnowledgeTip`'s own
- * per-source-user cap, same as a member's own `suggest_knowledge` calls.
- * Backed by the existing `answer_feedback_user_rate_idx (platform, user_id,
- * created_at DESC)`; the join to `interactions` is on its primary key and
- * scans at most `RATE_ANSWER_DAILY_LIMIT` rows.
- */
-export async function countMismatchedHelpfulRatings(
-  platform: Platform,
-  raterUserId: string,
-): Promise<number> {
-  const { rows } = await pool.query(
-    `SELECT count(*)::int AS n
-       FROM answer_feedback af
-       JOIN interactions i ON i.id = af.interaction_id
-      WHERE af.platform = $1 AND af.user_id = $2 AND af.helpful = true
-        AND af.created_at > now() - interval '24 hours'
-        AND i.meta->>'replyToUserId' IS NOT NULL
-        AND i.meta->>'replyToUserId' <> $2`,
-    [platform, raterUserId],
-  );
-  return rows[0]?.n ?? 0;
-}
-
-/**
- * Admin-tier view of answer feedback, scoped to `conversationIds` (null =
- * super admin, unrestricted — same convention as `listReports`). A rating
- * from a conversation no ordinary admin participates in is therefore only
- * reachable here with the unrestricted (super admin) scope.
- */
-export async function listAnswerFeedback(
-  conversationIds: readonly string[] | null,
-  unhelpfulOnly = false,
-  limit = 50,
-): Promise<AnswerFeedback[]> {
-  const params: unknown[] = [];
-  const filters: string[] = [];
-  if (conversationIds) {
-    params.push([...conversationIds]);
-    filters.push(`answer_feedback.conversation_id = ANY($${params.length})`);
-  }
-  if (unhelpfulOnly) {
-    filters.push(`answer_feedback.helpful = false`);
-  }
-  const clampedLimit = Math.min(Math.max(Math.trunc(limit) || 50, 1), 200);
-  params.push(clampedLimit);
-  const where = filters.length ? `WHERE ${filters.join(' AND ')}` : '';
-
-  // LEFT JOIN interactions to surface the rated answer's text and (when
-  // served via the deterministic knowledge shortcut) which knowledge entry
-  // produced it (issue #269) — both read through the SAME conversation_id
-  // scope filter above, so an admin outside the rated conversation never
-  // sees the enrichment either.
-  const { rows } = await pool.query(
-    `SELECT answer_feedback.id, answer_feedback.platform, answer_feedback.conversation_id,
-            answer_feedback.user_id, answer_feedback.interaction_id, answer_feedback.helpful,
-            answer_feedback.created_at, answer_feedback.comment, interactions.content,
-            (interactions.meta->>'knowledgeEntryId')::bigint AS knowledge_entry_id
-       FROM answer_feedback
-       LEFT JOIN interactions ON interactions.id = answer_feedback.interaction_id
-       ${where}
-      ORDER BY answer_feedback.created_at DESC
-      LIMIT $${params.length}`,
-    params,
-  );
-  return rows.map(mapAnswerFeedback);
-}
-
-export interface KnowledgeFeedbackSummary {
-  knowledgeEntryId: number;
-  title: string | null;
-  helpfulCount: number;
-  unhelpfulCount: number;
-  updatedAt: Date;
-  sampleComment: string | null;
-}
-
-function mapKnowledgeFeedbackSummary(r: {
-  id: number | string;
-  title: string | null;
-  updated_at: Date;
-  helpful_count: number | string;
-  unhelpful_count: number | string;
-  sample_comment: string | null;
-}): KnowledgeFeedbackSummary {
-  return {
-    knowledgeEntryId: Number(r.id),
-    title: r.title,
-    helpfulCount: Number(r.helpful_count),
-    unhelpfulCount: Number(r.unhelpful_count),
-    updatedAt: r.updated_at,
-    sampleComment: r.sample_comment,
-  };
-}
-
-/**
- * Admin-tier aggregation of `answer_feedback` per knowledge entry (issue
- * #287), the grouped complement to `listAnswerFeedback`'s flat per-row view.
- * Reuses the SAME `answer_feedback` → `interactions` join and
- * `conversation_id = ANY($1)` scope filter (null = super admin, unrestricted)
- * `listAnswerFeedback` already uses, so an admin never counts a rating from a
- * conversation they don't participate in. Ratings on interactions with no
- * `knowledgeEntryId` never join to a `knowledge` row and are therefore never
- * counted. `knowledgeEntryId` is written both by the deterministic knowledge
- * shortcut (exact match) and, since issue #411, best-effort by the normal
- * model-mediated `knowledge_search` path — a correlation with the most
- * recent qualifying hit in the turn, not a guarantee the model's reply
- * actually drew from that entry (see `AgentReply.knowledgeEntryId` in
- * `agent/core.ts`). Only entries with `unhelpfulCount >= minUnhelpful` are
- * returned, sorted by `unhelpfulCount` descending. `sampleComment` (issue
- * #409) is the most recent non-null `comment` (#355) from an *unhelpful*
- * rating on that entry, or null when none exists — comments on helpful
- * ratings are never selected, since they aren't signal for what's wrong with
- * the entry. Drawn from the same scope-filtered rows as the counts above, so
- * a comment from a conversation outside `conversationIds` can never surface.
- *
- * `interactions.created_at >= knowledge.updated_at` (issue #540) excludes
- * ratings on interactions that predate the entry's most recent
- * `update_knowledge` edit, so fixing a flagged entry resets its counts here
- * instead of the pre-edit unhelpful ratings counting against the new
- * content forever.
- */
-export async function listKnowledgeFeedbackSummary(
-  conversationIds: readonly string[] | null,
-  minUnhelpful = 2,
-  limit = 20,
-): Promise<KnowledgeFeedbackSummary[]> {
-  const params: unknown[] = [];
-  const filters: string[] = [
-    `(interactions.meta->>'knowledgeEntryId') IS NOT NULL`,
-    `interactions.created_at >= knowledge.updated_at`,
-  ];
-  if (conversationIds) {
-    params.push([...conversationIds]);
-    filters.push(`answer_feedback.conversation_id = ANY($${params.length})`);
-  }
-  params.push(Math.max(Math.trunc(minUnhelpful) || 2, 1));
-  const minUnhelpfulParam = params.length;
-  const clampedLimit = Math.min(Math.max(Math.trunc(limit) || 20, 1), 100);
-  params.push(clampedLimit);
-
-  const { rows } = await pool.query(
-    `SELECT knowledge.id, knowledge.title, knowledge.updated_at,
-            count(*) FILTER (WHERE answer_feedback.helpful) AS helpful_count,
-            count(*) FILTER (WHERE NOT answer_feedback.helpful) AS unhelpful_count,
-            (array_agg(answer_feedback.comment ORDER BY answer_feedback.created_at DESC)
-              FILTER (WHERE answer_feedback.comment IS NOT NULL AND NOT answer_feedback.helpful))[1]
-              AS sample_comment
-       FROM answer_feedback
-       JOIN interactions ON interactions.id = answer_feedback.interaction_id
-       JOIN knowledge ON knowledge.id = (interactions.meta->>'knowledgeEntryId')::bigint
-      WHERE ${filters.join(' AND ')}
-      GROUP BY knowledge.id, knowledge.title, knowledge.updated_at
-     HAVING count(*) FILTER (WHERE NOT answer_feedback.helpful) >= $${minUnhelpfulParam}
-      ORDER BY unhelpful_count DESC
-      LIMIT $${params.length}`,
-    params,
-  );
-  return rows.map(mapKnowledgeFeedbackSummary);
-}
-
-/**
- * Entry-scoped low-rated check for the member knowledge-shortcut serve path
- * (issue #337) — the same `answer_feedback` -> `interactions` join
- * `listKnowledgeFeedbackSummary` uses, narrowed to one entry id, but
- * deliberately UNSCOPED by conversation: there is no admin identity to scope
- * to at serve time (the caller is the member being served, not an admin
- * viewing their own conversations).
- *
- * SECURITY: returns only the threshold decision the SQL itself computes
- * (`>= $2`), never the raw unhelpful count or any per-rating row — the
- * caller-side render path must never see a number derived from the
- * aggregate, since `minUnhelpful` is enforced to be >= 2 specifically so no
- * single identifiable rater can be inferred from it (config.ts).
- *
- * `interactions.created_at >= knowledge.updated_at` (issue #540) excludes
- * ratings from before the entry's most recent `update_knowledge` edit, so a
- * fixed entry stops being reported low-rated once its pre-edit ratings are
- * the only ones on record.
- */
-export async function isKnowledgeLowRated(entryId: number, minUnhelpful: number): Promise<boolean> {
-  const { rows } = await pool.query(
-    `SELECT count(*) FILTER (WHERE NOT answer_feedback.helpful) >= $2 AS is_low_rated
-       FROM answer_feedback
-       JOIN interactions ON interactions.id = answer_feedback.interaction_id
-       JOIN knowledge ON knowledge.id = (interactions.meta->>'knowledgeEntryId')::bigint
-      WHERE (interactions.meta->>'knowledgeEntryId')::bigint = $1
-        AND interactions.created_at >= knowledge.updated_at`,
-    [entryId, minUnhelpful],
-  );
-  return rows[0]?.is_low_rated === true;
-}
-
-/**
- * Batched sibling of `isKnowledgeLowRated` (issue #432) — the normal
- * `knowledge_search` path checks many hits per call, so this exists to avoid
- * one query per hit; same join and `>= $2` threshold, but `ANY($1)` +
- * `GROUP BY` over a whole id list at once, returning only the subset that
- * crosses the threshold.
- *
- * SECURITY: same aggregate-only posture as `isKnowledgeLowRated` — the
- * returned `Set<number>` carries only which ids cleared the threshold, never
- * a raw unhelpful count or any per-rating row, preserving the "no single
- * identifiable rater can be inferred" property `minUnhelpful`'s `>= 2` floor
- * (config.ts) exists to protect.
- *
- * Short-circuits to an empty set for an empty `entryIds` array without
- * issuing a query — mirrors `hasConflictAmongIds`'s own zero-query
- * short-circuit for a too-small input.
- *
- * `interactions.created_at >= knowledge.updated_at` (issue #540) — same
- * post-edit reset as `isKnowledgeLowRated`.
- */
-export async function areKnowledgeEntriesLowRated(
-  entryIds: readonly number[],
-  minUnhelpful: number,
-): Promise<Set<number>> {
-  if (entryIds.length === 0) return new Set();
-  const { rows } = await pool.query(
-    `SELECT knowledge.id
-       FROM answer_feedback
-       JOIN interactions ON interactions.id = answer_feedback.interaction_id
-       JOIN knowledge ON knowledge.id = (interactions.meta->>'knowledgeEntryId')::bigint
-      WHERE knowledge.id = ANY($1)
-        AND interactions.created_at >= knowledge.updated_at
-      GROUP BY knowledge.id
-     HAVING count(*) FILTER (WHERE NOT answer_feedback.helpful) >= $2`,
-    [entryIds, minUnhelpful],
-  );
-  return new Set(rows.map((r) => Number(r.id)));
-}
-
-/**
- * Count distinct knowledge entries with `unhelpfulCount >= minUnhelpful`
- * (issue #324), for the weekly admin digest — the growth path #287 itself
- * named. Reuses the SAME `answer_feedback` → `interactions` → `knowledge`
- * join, scope filter, and `HAVING` clause as `listKnowledgeFeedbackSummary`,
- * but a true `SELECT count(DISTINCT ...)`, never `.length` of that
- * function's `LIMIT`-bounded list, so a backlog past its default `limit` of
- * 20 is not understated. **Conversation-scoped** like `countKnowledgeGaps`/
- * `countOpenReports` (null = super admin, unrestricted) — an admin never
- * counts a rating from a conversation they don't participate in. Ratings on
- * interactions with no `knowledgeEntryId` never join to a `knowledge` row
- * and are therefore never counted, matching `listKnowledgeFeedbackSummary`'s
- * existing boundary.
- *
- * `interactions.created_at >= knowledge.updated_at` (issue #540) — same
- * post-edit reset as `listKnowledgeFeedbackSummary`, so this count and that
- * list agree after an admin fixes an entry.
- */
-export async function countLowRatedKnowledge(
-  conversationIds: readonly string[] | null,
-  minUnhelpful = 2,
-): Promise<number> {
-  const params: unknown[] = [];
-  const filters: string[] = [
-    `(interactions.meta->>'knowledgeEntryId') IS NOT NULL`,
-    `interactions.created_at >= knowledge.updated_at`,
-  ];
-  if (conversationIds) {
-    params.push([...conversationIds]);
-    filters.push(`answer_feedback.conversation_id = ANY($${params.length})`);
-  }
-  params.push(Math.max(Math.trunc(minUnhelpful) || 2, 1));
-  const minUnhelpfulParam = params.length;
-
-  const { rows } = await pool.query(
-    `SELECT count(*) AS n FROM (
-       SELECT knowledge.id
-         FROM answer_feedback
-         JOIN interactions ON interactions.id = answer_feedback.interaction_id
-         JOIN knowledge ON knowledge.id = (interactions.meta->>'knowledgeEntryId')::bigint
-        WHERE ${filters.join(' AND ')}
-        GROUP BY knowledge.id
-       HAVING count(*) FILTER (WHERE NOT answer_feedback.helpful) >= $${minUnhelpfulParam}
-     ) low_rated`,
-    params,
-  );
-  return Number(rows[0].n);
-}
-
-export interface AnswerFeedbackOriginSummary {
-  autoAnswer: { helpful: number; unhelpful: number };
-  addressed: { helpful: number; unhelpful: number };
-}
-
-/**
- * Guild-wide (by `conversationIds` scope) split of `answer_feedback` ratings
- * by delivery origin — auto-answered (issue #477, unsummoned) vs addressed
- * (@mention/DM/thread-reply) — the answer-**quality** counterpart to
- * `usageStats`'s `autoAnswerUsage` **cost** split (issue #552), which reads
- * the SAME `meta.autoAnswer` flag from the opposite side (a plain SUM/COUNT
- * over `interactions` alone, vs this one joining through `answer_feedback`).
- * Reuses the identical `answer_feedback JOIN interactions` join and
- * `count(*) FILTER (WHERE ...)` aggregate shape `listKnowledgeFeedbackSummary`/
- * `countLowRatedKnowledge` already use — no schema change, no migration.
- *
- * Cumulative (no freshness window), matching `countLowRatedKnowledge`'s own
- * cumulative-backlog shape rather than a rolling-window count — a helpful
- * ratio is more meaningful over the whole history than reset weekly against a
- * handful of ratings.
- *
- * `conversationIds` scoping mirrors every other admin-tier aggregate here
- * (`listAnswerFeedback`/`listKnowledgeFeedbackSummary`/`countLowRatedKnowledge`):
- * `null` means unrestricted (super admin); otherwise only ratings whose
- * `answer_feedback.conversation_id` is in the given list are counted, so an
- * admin never counts a rating from a conversation they don't participate in.
- *
- * SECURITY: bucketing is driven SOLELY by `interactions.meta->>'autoAnswer'`
- * — router-internal state set only from `replyConversationId !== undefined`
- * (issue #552/#553), never from message content — so a rated reply whose
- * text happens to resemble the flag can never move it between buckets. A
- * rating whose `interaction_id` was cleared (`ON DELETE SET NULL`, e.g. after
- * `forget_me`/`purge_user_data`) has no interaction row left to join against
- * and is excluded from BOTH buckets, matching `listKnowledgeFeedbackSummary`'s
- * own INNER JOIN exclusion of a purged interaction.
- */
-export async function answerFeedbackOriginSummary(
-  conversationIds: readonly string[] | null,
-): Promise<AnswerFeedbackOriginSummary> {
-  const params: unknown[] = [];
-  const filters: string[] = [];
-  if (conversationIds) {
-    params.push([...conversationIds]);
-    filters.push(`answer_feedback.conversation_id = ANY($${params.length})`);
-  }
-  const where = filters.length ? `WHERE ${filters.join(' AND ')}` : '';
-
-  const { rows } = await pool.query(
-    `SELECT
-       count(*) FILTER (WHERE (interactions.meta->>'autoAnswer') IS NOT NULL AND answer_feedback.helpful)
-         AS auto_helpful,
-       count(*) FILTER (WHERE (interactions.meta->>'autoAnswer') IS NOT NULL AND NOT answer_feedback.helpful)
-         AS auto_unhelpful,
-       count(*) FILTER (WHERE (interactions.meta->>'autoAnswer') IS NULL AND answer_feedback.helpful)
-         AS addressed_helpful,
-       count(*) FILTER (WHERE (interactions.meta->>'autoAnswer') IS NULL AND NOT answer_feedback.helpful)
-         AS addressed_unhelpful
-       FROM answer_feedback
-       JOIN interactions ON interactions.id = answer_feedback.interaction_id
-       ${where}`,
-    params,
-  );
-  const r = rows[0];
-  return {
-    autoAnswer: { helpful: Number(r.auto_helpful), unhelpful: Number(r.auto_unhelpful) },
-    addressed: { helpful: Number(r.addressed_helpful), unhelpful: Number(r.addressed_unhelpful) },
-  };
 }
